Finalização do teste da classe Curso e instalação do  Bogus
