+++ src/CursoOnline.Dominio/Cursos/PublicoAlvo.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace CursoOnline.Dominio.Cursos
+{
+    public enum PublicoAlvo
+    {
+        Estudante,
+        Universitario,
+        Empregado,
+        Empreendedor
+    }
+}
